Rework Go Tests (#1228)
