+++ backup/providers_test.go
@@ -97,10 +97,10 @@ func TestAzure_Init(t *testing.T) {
 		{
 			"Success",
 			fields{
-				AzureStorageAccount: help.GetEnv("AZURE_STORAGE_ACCOUNT", ""),
-				AzureStorageKey:     help.GetEnv("AZURE_STORAGE_TOKEN", ""),
+				AzureStorageAccount: help.GetEnv("AZURE_ACCOUNT_NAME", ""),
+				AzureStorageKey:     help.GetEnv("AZURE_ACCOUNT_KEY", ""),
 				AzureContainerName:  help.GetEnv("AZURE_CONTAINER_NAME", ""),
-				AzureStoragePath:    help.GetEnv("AZURE_STORAGE_PATH", ""),
+				AzureStoragePath:    help.GetEnv("AZURE_BACKUP_STORAGE_PATH", ""),
 			},
 			false,
 		},
@@ -123,9 +123,7 @@ func TestAzure_Init(t *testing.T) {
 
 // TODO: convert into file-store agnostic test and add cases for other file stores
 func TestAzure_Upload(t *testing.T) {
-	if testing.Short() {
-		t.Skip()
-	}
+	t.Skip("not currently supported")
 	_ = godotenv.Load(".env")
 	type fields struct {
 		AzureStorageAccount string
@@ -183,9 +181,7 @@ func TestAzure_Upload(t *testing.T) {
 }
 
 func TestAzure_Download(t *testing.T) {
-	if testing.Short() {
-		t.Skip()
-	}
+	t.Skip("not currently supported")
 	_ = godotenv.Load(".env")
 	type fields struct {
 		AzureStorageAccount string
