+++ codecov.yml
@@ -1,13 +1,14 @@
-github_checks: false
+github_checks: true
 coverage:
   status:
     project:
       default:
-        target: 0
+        target: auto
     patch:
       default:
-        target: 0
+        target: 75%
 codecov:
+  branch: main
   require_ci_to_pass: true
   notify:
     wait_for_ci: true
