+++ helpers/compression/compression_test.go
@@ -1,6 +1,3 @@
-//go:build compress
-// +build compress
-
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
