+++ provider/scripts/k8s/requirements.txt
@@ -1,8 +1,8 @@
 dill==0.3.7
 pandas>=1.3.5,<2.2.0
 pandasql==0.7.3
-pyarrow==14.0.1
-fastparquet==0.8.3
+pyarrow
+fastparquet
 pytest==7.1.2
 pytest-cov==3.0.0
 python-etcd==0.4.5
