+++ tests/integration/backup/load_test.go
@@ -1,6 +1,3 @@
-//go:build backup
-// +build backup
-
 package main
 
 import (
@@ -12,6 +9,7 @@ import (
 )
 
 func TestReadETCD(t *testing.T) {
+	t.Skip("unmaintained")
 	host := help.GetEnv("ETCD_HOST", "localhost")
 	port := help.GetEnv("ETCD_PORT", "2379")
 	address := fmt.Sprintf("%s:%s", host, port)
